Update 230727
Updated AddItemToCartController
Updated all controllers that use this function to work properly
Updated login and register controller
Updated route service provider, so it has a new route: LOGIN

PS: Now you can put stuff in the cart, register, sign in, and the cart stays.
I solved this 'problem' by deleting the cart associated with the user's ID, if the cart with the unauthenticated user (so with session id only) has items. If it as no items, that cart will be deleted, and the one with the user ID stays.