Update 230801 v2
Updated migrations
Updated vue components. They're now a bit cleaner but it's a clear mindfuck.
Updated Cart and CartItems model functions.
Updated DeleteOldCarts job so it uses the model function to delete carts.
Updated the cart controller so we validate the product_id on removeItem before we actually remove it.