Update 230826
Updated controllers
Updated AddItemToCartController, so it throws errors more approprietly. I'll update the rest of the controllers for the same error handling.
Updated the custom emittable events in app.js, so they're able to handle custom callback functions.
In order for this code to be used in *any* emittable events, I created the HelperFunctions.js file, with the handleCallback function.
Also updated some vue components to make use of this custom callback functionality.